% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire

A data set collected concerning home damage from fire}
\alias{fire

A data set collected concerning home damage from fire}
\alias{fire}
\title{Fire damage}
\format{
## `fire`
A data frame with 15 rows and 2 variables
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
Fire damage
}
\keyword{datasets}
