% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Calculate Confidence Interval for the Mean}
\usage{
myci(x)
}
\arguments{
\item{x}{Numeric vector of data.}
}
\value{
A numeric vector representing the confidence interval for the mean.
}
\description{
Calculate Confidence Interval for the Mean
}
\examples{

x = rnorm(30,mean=10,sd=12)
}
