% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project1.R
\name{myddt}
\alias{myddt}
\title{Project 1 function}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{data frame}

\item{SPECIES}{the species you wish to filter}
}
\value{
plot and csv file plus data and table
}
\description{
This is the function needed for project 1 of MATH 4753.
}
\examples{
\dontrun{obj <- myddt(df = ddt, SPECIES = "CCATFISH")}
}
