% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Make a normal}
\usage{
myf(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations to simulate.}

\item{n}{Sample size for each iteration.}

\item{p}{Probability of success for each trial.}
}
\value{
A histogram
}
\description{
This function will make a plot of a random sample
}
