% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{myfunNSE}
\alias{myfunNSE}
\title{filtering data}
\usage{
myfunNSE(df, cond)
}
\arguments{
\item{df}{data frame}

\item{cond}{condition for filter}
}
\value{
filtered data frame
}
\description{
filtering data
}
\examples{
myfunNSE(ddt, cond = LENGTH > 40)
}
