% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mu for the dnorm}

\item{sigma}{sigma for the dnorm}

\item{a}{x value for shaded area}
}
\value{
curve
}
\description{
myncurve
}
\examples{
\dontrun{myncurve(mu,sigma,a)}
}
