% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Project 1 function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of tickets.}

\item{gamma}{The gamma parameter for the calculation.}

\item{p}{The probability parameter for the calculation.}
}
\value{
The result of the calculation.
}
\description{
Project 1 function
}
\examples{
ntickets(N=400, gamma = 0.02, p = 0.95)
}
