% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Ticket_Approximation}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of total overbooking}

\item{p}{probability of show up}
}
\description{
Ticket_Approximation
}
\examples{
\dontrun{ntickets(200,0.02,0.95)}

}
